Move `rray_clip()` to its own file
