+++ man/rray_clip.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/math-basic.R
+% Please edit documentation in R/clip.R
 \name{rray_clip}
 \alias{rray_clip}
 \title{Bound the values of an array}
@@ -25,8 +25,3 @@ x <- matrix(1:10, ncol = 2)
 rray_clip(x, 1, 5)
 
 }
-\seealso{
-Other math functions: \code{\link{rray_maximum}},
-  \code{\link{rray_multiply_add}}
-}
-\concept{math functions}
